+++ _com/Estilos.css
@@ -1,77 +1,6 @@
-
-@import url("https://fonts.googleapis.com/css?family=Varela+Round");
-
-
-nav{
-    width: 100%;
-}
-
-.menu {
-    width: 100%;
-    height: 100%;
-    background: #1c1d22;
-    font-family: "Varela Round", sans-serif;
-}
-
-.menu {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    max-width: 100%;
-    margin: 0 auto;
-    height: 70px;
-    list-style: none;
-}
-
-.menu li {
-    width: 125px;
-    height: 50px;
-    transition: background-position-x 0.9s linear;
-    text-align: center;
-}
-.menu li a {
-    font-size: 22px;
-    color: #f6f5f5;
-    text-decoration: none;
-    transition: all 0.45s;
-}
-.menu li:hover {
-    background: url("data:image/svg+xml;base64,PD94bWwgdmVyc2lvbj0iMS4wIiBlbmNvZGluZz0idXRmLTgiPz4KPHN2ZyB2ZXJzaW9uPSIxLjEi%0D%0AIGlkPSJMYXllcl8xIiB4bWxucz0iaHR0cDovL3d3dy53My5vcmcvMjAwMC9zdmciIHhtbG5zOnhs%0D%0AaW5rPSJodHRwOi8vd3d3LnczLm9yZy8xOTk5L3hsaW5rIiB4PSIwcHgiIHk9IjBweCIKCSB3aWR0%0D%0AaD0iMzkwcHgiIGhlaWdodD0iNTBweCIgdmlld0JveD0iMCAwIDM5MCA1MCIgZW5hYmxlLWJhY2tn%0D%0Acm91bmQ9Im5ldyAwIDAgMzkwIDUwIiB4bWw6c3BhY2U9InByZXNlcnZlIj4KPHBhdGggZmlsbD0i%0D%0Abm9uZSIgc3Ryb2tlPSIjZDk0ZjVjIiBzdHJva2Utd2lkdGg9IjEuNSIgc3Ryb2tlLW1pdGVybGlt%0D%0AaXQ9IjEwIiBkPSJNMCw0Ny41ODVjMCwwLDk3LjUsMCwxMzAsMAoJYzEzLjc1LDAsMjguNzQtMzgu%0D%0ANzc4LDQ2LjE2OC0xOS40MTZDMTkyLjY2OSw0Ni41LDI0My42MDMsNDcuNTg1LDI2MCw0Ny41ODVj%0D%0AMzEuODIxLDAsMTMwLDAsMTMwLDAiLz4KPC9zdmc+Cg==");
-    -webkit-animation: line 1s;
-    animation: line 1s;
-}
-.menu li:hover a {
-    color: #d94f5c;
-}
-.menu li:not(:last-child) {
-    margin-right: 20%;
-}
-
-@-webkit-keyframes line {
-    0% {
-        background-position-x: 390px;
-    }
-}
-
-@keyframes line {
-    0% {
-        background-position-x: 390px;
-    }
-}
-
-
-
-
-
-
-
-h1 {
-    text-align: center;
-}
-
 #secTabla > div {
     display: table;
-    width: 100%;
+    width: 40%;
 }
 
 #secTabla > div > div {
@@ -90,10 +19,9 @@ h1 {
 #secTabla > div > div > div > input {
     width: 100%;
     border: none;
-    text-align: center;
 }
 
-#crearProducto, #crearUsuario{
+#crearProducto{
     border: 1px solid silver;
     width: 20%;
     padding: 5px;
@@ -132,4 +60,3 @@ h1 {
     border-color: black;
     border-radius: 20px;
 }
-
