+++ sesiones/SesionFormulario.php
@@ -32,7 +32,26 @@
     <input type='submit' value='Iniciar sesión'>
 </form>
 
-<h10>¿Nuevo en Tienda? <a href="crearUsuario.php" style="text-decoration: none">Crear una cuenta</a> Tienda</h10>
+
+<form action="UsuarioCrear.php" id="form" method="post">
+    <label for="nombre">Nombre</label> <br>
+    <input type="text" name="inpNombreUsuario"> <br> <br>
+
+    <label for="apellidos">Apellidos</label> <br>
+    <input type="text" name="inpApellidosUsuario"> <br> <br>
+
+    <label for="identificador">Identificador</label> <br>
+    <input type="text" name="inpIdentificadorUsuario"> <br> <br>
+
+    <label for="contrasenna">Contraseña</label> <br>
+    <input type="text" name="inpContrasennaUsuario"> <br> <br>
+
+    <label for="contrasenna2">Confirme su contraseña</label> <br>
+    <input type="text" name="inpContrasenna2Usuario"> <br> <br>
+
+    <input type="submit" name="btnRegistroUsuario">
+</form>
+
 </body>
 
 </html>
